# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ultimate Life OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /habits:
    get:
      operationId: getHabits
      tags: [habits]
      summary: List all habits
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [habits]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitInput"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{id}:
    patch:
      operationId: updateHabit
      tags: [habits]
      summary: Update a habit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitUpdate"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [habits]
      summary: Delete a habit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /habits/{id}/log:
    post:
      operationId: logHabit
      tags: [habits]
      summary: Log a habit completion for today
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HabitLogInput"
      responses:
        "201":
          description: Created habit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitLog"
  /habits/{id}/logs:
    get:
      operationId: getHabitLogs
      tags: [habits]
      summary: Get logs for a habit (last 30 days)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Habit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabitLog"
...
  /goals:
    get:
      operationId: getGoals
      tags: [goals]
      summary: List all goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a goal by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Goal with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalWithSteps"
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /goals/{id}/steps:
    post:
      operationId: createGoalStep
      tags: [goals]
      summary: Add a step to a goal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalStepInput"
      responses:
        "201":
          description: Created goal step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalStep"
  /goals/{id}/steps/{stepId}:
    patch:
      operationId: updateGoalStep
      tags: [goals]
      summary: Update a goal step
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: stepId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalStepUpdate"
      responses:
        "200":
          description: Updated goal step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalStep"

  # ─── STATISTICS ──────────────────────────────────────────────────────────────
...
  /ai/daily-report:
    get:
      operationId: getAiDailyReport
      tags: [ai]
      summary: Get AI-generated daily analysis
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
      responses:
        "200":
          description: AI daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReport"
  /ai/messages:
    get:
      operationId: getAiMessages
      tags: [ai]
      summary: Get AI messages/notifications
      responses:
        "200":
          description: AI messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMessage"
    post:
      operationId: sendAiMessage
      tags: [ai]
      summary: Send message to AI assistant and get response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessage"
components:
  schemas:
...
    Habit:
      type: object
      required: [id, name, icon, currentStreak, bestStreak, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        targetDays:
          type: string
          enum: [daily, weekdays, weekends, custom]
        currentStreak:
          type: integer
        bestStreak:
          type: integer
        totalCompletions:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    HabitInput:
      type: object
      required: [name, icon]
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
        color:
          type: string
        description:
          type: string
        targetDays:
          type: string
          enum: [daily, weekdays, weekends, custom]
    HabitUpdate:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        description:
          type: string
        targetDays:
          type: string
          enum: [daily, weekdays, weekends, custom]
        isActive:
          type: boolean
    HabitLog:
      type: object
      required: [id, habitId, date, completed]
      properties:
        id:
          type: integer
        habitId:
          type: integer
        date:
          type: string
          format: date
        completed:
          type: boolean
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    HabitLogInput:
      type: object
      required: [date, completed]
      properties:
        date:
          type: string
          format: date
        completed:
          type: boolean
        note:
          type: string
...
    Goal:
      type: object
      required: [id, title, progressPercent, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
          format: date
        progressPercent:
          type: number
        status:
          type: string
          enum: [active, completed, paused, abandoned]
        priority:
          type: string
          enum: [low, medium, high]
        totalSteps:
          type: integer
        completedSteps:
          type: integer
        aiAdvice:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GoalInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        category:
          type: string
        deadline:
          type: string
          format: date
        priority:
          type: string
          enum: [low, medium, high]
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        deadline:
          type: string
          format: date
        status:
          type: string
          enum: [active, completed, paused, abandoned]
        priority:
          type: string
          enum: [low, medium, high]
        progressPercent:
          type: number
    GoalStep:
      type: object
      required: [id, goalId, title, completed, order]
      properties:
        id:
          type: integer
        goalId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        completed:
          type: boolean
        order:
          type: integer
        deadline:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    GoalStepInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        deadline:
          type: string
          format: date
        order:
          type: integer
    GoalStepUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
        deadline:
          type: string
          format: date
        order:
          type: integer
    GoalWithSteps:
      type: object
      required: [goal, steps]
      properties:
        goal:
          $ref: "#/components/schemas/Goal"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/GoalStep"

    # ─── STATS SCHEMAS ──────────────────────────────────────────────────────────
...
    AiReport:
      type: object
      required: [date, analysis, messages, recommendation, mood, productivityScore]
      properties:
        date:
          type: string
          format: date
        analysis:
          type: string
        messages:
          type: array
          items:
            type: string
        recommendation:
          type: string
        mood:
          type: string
          enum: [excellent, good, average, poor, critical]
        productivityScore:
          type: number
        warnings:
          type: array
          items:
            type: string
        compliments:
          type: array
          items:
            type: string
    AiMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        messageType:
          type: string
          enum: [chat, daily_report, warning, encouragement, advice]
    AiMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1