# lib/api-zod/src/generated/types/aiMessageInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ultimate Life OS API
 * OpenAPI spec version: 0.1.0
 */

export interface AiMessageInput {
  /** @minLength 1 */
  content: string;
}