# lib/api-zod/src/generated/types/goalStepInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Ultimate Life OS API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalStepInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  deadline?: Date;
  order?: number;
}